feat: always specify definition file as argument
This opens up for future usage where we have multiple definition
files for different contexts. It also makes it more explicit for
the users that they have to keep the file up-to-date.

BREAKING CHANGE: --definition-file must now be specified for
some of the CLI commands, and how the definition file is loaded
is changed so it no longer reads from the config file or environment
variable CALS_DEFINITION_FILE.