feat: remove RepoDeleteItem
We don't have a standardized way of creating repos, and they will often
exist independently of resources.yaml. They may be created before, or
they may be a short-lived PoC repo. Either way, we don't want to delete
them just because they're untracked.

This change removes the repo-delete as a recognized type. Both because
we don't want to perform deletion, and because it floods the planned
change output with messages like

```
info   -
{"type":"repo-delete","org":"capralifecycle","repo":"tono-infra"}
info   -
{"type":"repo-delete","org":"capralifecycle","repo":"tono-music-monitor-poc"}
info   -
{"type":"repo-delete","org":"capralifecycle","repo":"tono-performances-web"}
info   -
{"type":"repo-delete","org":"capralifecycle","repo":"tono-sync"}
info   - {"type":"repo-delete","org":"capralifecycle","repo":"tono-tgf"}
info   - {"type":"repo-delete","org":"capralifecycle","repo":"tono-tix"}
info   -
{"type":"repo-delete","org":"capralifecycle","repo":"tono-typelib"}
info No actions to be performed
```

If we want to do something with unrecognized repos in the future, we can
revert this change, but until then, I suggest we clean it up